// src/app/openapi/model/catalog.ts
/**
 * Component Catalog REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CatalogLink } from './catalog-link';


export interface Catalog { 
    name?: string;
    description?: string;
    communityPageId?: string;
    links?: Array<CatalogLink>;
    tags?: Array<string>;
}