// src/app/openapi/model/file-format.ts
/**
 * Component Catalog REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type FileFormat = 'image' | 'markdown' | 'yaml';

export const FileFormat = {
    Image: 'image' as FileFormat,
    Markdown: 'markdown' as FileFormat,
    Yaml: 'yaml' as FileFormat
};